+++ pool/app/tags/tags.ml
@@ -2,3 +2,7 @@ include Entity
 include Event
 include Repo
 module Guard = Entity_guard
+
+module ParticipationTags = struct
+  include Repo_participation_tags
+end