+++ pool/app/tags/tags.mli
@@ -59,11 +59,23 @@ val create
   -> Model.t
   -> (t, Pool_common.Message.error) result
 
+module ParticipationTags : sig
+  type entity =
+    | Experiment of Pool_common.Id.t
+    | Session of Pool_common.Id.t
+
+  val get_id : entity -> Pool_common.Id.t
+  val find_all : Pool_database.Label.t -> entity -> t list Lwt.t
+  val find_available : Pool_database.Label.t -> entity -> t list Lwt.t
+end
+
 type event =
   | Created of t
   | Updated of t
   | Tagged of Tagged.t
   | Untagged of Tagged.t
+  | ParticipationTagAssigned of ParticipationTags.entity * Id.t
+  | ParticipationTagRemoved of ParticipationTags.entity * Id.t
 
 val equal_event : event -> event -> bool
 val pp_event : Format.formatter -> event -> unit
@@ -160,4 +172,20 @@ module Guard : sig
     val assign : ('a -> Guard.ValidationSet.t) -> 'a -> Guard.ValidationSet.t
     val remove : ('a -> Guard.ValidationSet.t) -> 'a -> Guard.ValidationSet.t
   end
+end
+
+module RepoEntity : sig
+  module Id : sig
+    val t : Id.t Caqti_type.t
+  end
+
+  module Model : sig
+    val t : Model.t Caqti_type.t
+  end
+
+  val t : t Caqti_type.t
+end
+
+module Sql : sig
+  val select_tag_sql : string
 end