+++ pool/database/migration/migration.ml
@@ -74,6 +74,7 @@ module Root = struct
       ; Migration_202306021033.migration ()
       ; Migration_202306261642.migration_root ()
       ; Migration_202307121722.migration_root ()
+      ; Migration_202307131619.migration_root ()
       ]
   ;;
 
@@ -122,6 +123,7 @@ module Tenant = struct
       ; Migration_202307121330.migration ()
       ; Migration_202307121722.migration ()
       ; Migration_202307131619.migration ()
+      ; Migration_202307250839.migration ()
       ]
   ;;