+++ pool/test/tag_test.ml
@@ -208,3 +208,47 @@ let try_assign_experiment_tag_to_contact _ () =
   in
   Lwt.return_unit
 ;;
+
+let assign_auto_tag_to_experiment () =
+  let open CCResult in
+  let open Tags in
+  let experiment = Test_utils.Model.create_experiment () in
+  let experiment_id = Experiment.(Id.to_common experiment.id) in
+  let tag = Data.Tag.create_with_description () in
+  let events =
+    let open Cqrs_command.Tags_command.AssignParticipationTagToEntity in
+    Pool_common.Message.[ Field.(show Tag), [ Tags.(Id.value tag.id) ] ]
+    |> decode
+    >>= handle (ParticipationTags.Experiment experiment_id)
+  in
+  let expected =
+    Ok
+      [ Tags.(
+          ParticipationTagAssigned
+            (ParticipationTags.Experiment experiment_id, tag.id))
+        |> Pool_event.tags
+      ]
+  in
+  Test_utils.check_result expected events
+;;
+
+let remove_auto_tag_from_experiment () =
+  let open CCResult in
+  let open Tags in
+  let experiment = Test_utils.Model.create_experiment () in
+  let experiment_id = Experiment.(Id.to_common experiment.id) in
+  let tag = Data.Tag.create_with_description () in
+  let events =
+    let open Cqrs_command.Tags_command.RemoveParticipationTagFromEntity in
+    handle (ParticipationTags.Experiment experiment_id) tag
+  in
+  let expected =
+    Ok
+      [ Tags.(
+          ParticipationTagRemoved
+            (ParticipationTags.Experiment experiment_id, tag.id))
+        |> Pool_event.tags
+      ]
+  in
+  Test_utils.check_result expected events
+;;